fix: skipping torrc backup when for system copy
The wizard attempts to back up the torrc but, if we're editing the system copy,
we'll lack the permissions for this. The override.py makes its own backup so
this is fine.

+++ src/cli/wizard.py
@@ -353,7 +353,7 @@ def showWizard():
           
           # if the torrc already exists then save it to a _bak file
           isBackedUp = False
-          if os.path.exists(torrcLocation):
+          if os.path.exists(torrcLocation) and not isSystemReplace:
             try:
               shutil.copy(torrcLocation, torrcLocation + "_bak")
               isBackedUp = True
